Update patch set 5
Patch Set 5: Code-Review-1

The current Exynos BSP was designed for 4.4, so at a time where likely nobody thought about VoLTE on Android.
If you want to implement it, you might need to take a look at the QC HAL and replicated that on our SLSI repos or come up with your own logic.

Patch-set: 5